moved request and response structs to user.handler.go
