--- cmd/api/handlers/reqres/users.go
@@ -1,61 +1,0 @@
-package reqres
-
-import (
-	"time"
-
-	"go-lms-of-pupilfirst/cmd/models"
-
-	"github.com/pborman/uuid"
-	"github.com/pkg/errors"
-	"golang.org/x/crypto/bcrypt"
-)
-
-// UserLoginRequest spec for login request
-type UserLoginRequest struct {
-	Email    string `json:"email" validate:"required,email,unique"`
-	Password string `json:"password" validate:"required"`
-}
-
-// UserCreateRequest spec for signup request
-type UserCreateRequest struct {
-	Name            string     `json:"name" validate:"required" example:"Groot"`
-	Email           string     `json:"email" validate:"required,email,unique" example:"groot@golms.com"`
-	Password        string     `json:"password" validate:"required" example:"GrootSecret"`
-	PasswordConfirm string     `json:"password_confirm" validate:"required,eqfield=password" example:"GrootSecret"`
-	TimeZone        *time.Time `json:"timezone" validate:"required" example:"America/Anchorage"`
-}
-
-// ToUser converts UserCreateRequest to User object
-func (userCreateRequest *UserCreateRequest) ToUser() (*models.User, error) {
-	if userCreateRequest == nil {
-		return nil, errors.New("Null User Create Request")
-	}
-
-	passwordSalt := uuid.NewRandom().String()
-	saltedPassword := userCreateRequest.Password + passwordSalt
-	passwordHash, err := bcrypt.GenerateFromPassword([]byte(saltedPassword), bcrypt.DefaultCost)
-	if err != nil {
-		return nil, errors.Wrap(err, "Error generating password hash")
-	}
-
-	user := &models.User{
-		Name:         userCreateRequest.Name,
-		Email:        userCreateRequest.Email,
-		PasswordSalt: passwordSalt,
-		PasswordHash: passwordHash,
-		TimeZone:     userCreateRequest.TimeZone,
-	}
-	return user, nil
-}
-
-// UserInfoUpdateRequest - spec for updating user info
-type UserInfoUpdateRequest struct {
-	ID        string `json:"id" validate:"required,uuid" example:"c01bdef7-173f-4d29-3edc60baf6a2"`
-	Name      string `json:"name" validate:"min=3,max=10,omitempty"`
-	Phone     string `json:"phone" validate:"omitempty"`
-	Title     string `json:"title" validate:"omitempty"`
-	KeySkills string `json:"key_skills" validate:"omitempty"`
-	About     string `gorm:"type:text" json:"about" validate:"omitempty"`
-
-	TimeZone *time.Time `json:"timezone" validation:"omitempty"`
-}
